- introduce function barrier to remove type instability in `transfer_solution` associated with `ip_field::Interpolation`
- make `FerriteViz.postprocess` type stable
- remove unnecessary type annotations

| Commit  | transfer_solution             | solutionplot                |
|---------|-------------------------------|-----------------------------|
| master | 4.277s    0.663 GiB           | 53.078s  30.38GB|
| pr | 2.913s 0.439 GiB | 51.152s 29.94 GB |

The internal "kernel" for the transfer `_transfer_solution` is close to type grounded